flow: output metric
