+++ conf/service_conf.yaml
@@ -1,26 +1,9 @@
 use_registry: false
-use_deserialize_safe_module: false
 dependent_distribution: false
 encrypt_password: false
 encrypt_module: fate_arch.common.encrypt_utils#pwdecrypt
 private_key:
-party_id:
-hook_module:
-  client_authentication: fate_flow.hook.flow.client_authentication
-  site_authentication: fate_flow.hook.flow.site_authentication
-  permission: fate_flow.hook.flow.permission
-hook_server_name:
-authentication:
-  client:
-    switch: false
-    http_app_key:
-    http_secret_key:
-  site:
-    switch: false
-permission:
-  switch: false
-  component: false
-  dataset: false
+party_id: "10000"
 fateflow:
   # you must set real ip address, 127.0.0.1 and 0.0.0.0 is not supported
   host: 127.0.0.1
@@ -34,19 +17,6 @@ fateflow:
     grpc_port:
   # use random instance_id instead of {host}:{http_port}
   random_instance_id: false
-
-  # support rollsite/nginx/fateflow as a coordination proxy
-  # rollsite support fate on eggroll, use grpc protocol
-  # nginx support fate on eggroll and fate on spark, use http or grpc protocol, default is http
-  # fateflow support fate on eggroll and fate on spark, use http protocol, but not support exchange network mode
-
-  # format(proxy: rollsite) means rollsite use the rollsite configuration of fate_one_eggroll and nginx use the nginx configuration of fate_one_spark
-  # you also can customize the config like this(set fateflow of the opposite party as proxy):
-  # proxy:
-  #   name: fateflow
-  #   host: xx
-  #   http_port: xx
-  #   grpc_port: xx
   proxy: rollsite
   # support default/http/grpc
   protocol: default
@@ -86,22 +56,12 @@ fate_on_spark:
     home:
     cores_per_node: 20
     nodes: 2
-  linkis_spark:
-    cores_per_node: 20
-    nodes: 2
-    host: 127.0.0.1
-    port: 9001
-    token_code: MLSS
-    python_path: /data/projects/fate/python
   hive:
     host: 127.0.0.1
     port: 10000
     auth_mechanism:
     username:
     password:
-  linkis_hive:
-    host: 127.0.0.1
-    port: 9001
   hdfs:
     name_node: hdfs://fate-cluster
     # default /
@@ -135,35 +95,9 @@ fate_on_spark:
     host: 127.0.0.1
     http_port: 9300
     grpc_port: 9310
-# external services
 fateboard:
   host: 127.0.0.1
   port: 8080
-
-enable_model_store: false
-model_store_address:
-  # use mysql as the model store engine
-#  storage: mysql
-#  database: fate_model
-#  user: fate
-#  password: fate
-#  host: 127.0.0.1
-#  port: 3306
-  # other optional configs send to the engine
-#  max_connections: 10
-#  stale_timeout: 10
-
-  # use tencent cos as model store engine
-  storage: tencent_cos
-  Region:
-  SecretId:
-  SecretKey:
-  Bucket:
-
 servings:
   hosts:
     - 127.0.0.1:8000
-fatemanager:
-  host: 127.0.0.1
-  port: 8001
-  federatedId: 0
