+++ python/fate_flow/apps/scheduler/scheduler_app.py
@@ -1,6 +1,6 @@
 from webargs import fields
 
-from fate_flow.entity.dag_structures import DAGSchema
+from fate_flow.hub.parser.default import DAGSchema
 from fate_flow.operation.job_saver import ScheduleJobSaver
 from fate_flow.scheduler.job_scheduler import DAGScheduler
 from fate_flow.utils.api_utils import get_json_result, validate_request_json, task_request_json
