--- python/fate_flow/entity/_dag_parameters.py
@@ -1,41 +1,0 @@
-#
-#  Copyright 2019 The FATE Authors. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-#
-from ._base import BaseEntity
-
-
-class DagParameters(BaseEntity):
-    def __init__(self, **kwargs):
-        self.scheduler_party_id = None
-        self.initiator_party_id = None
-        self.stage = None
-        self.parties = None
-        self.conf = {}
-        self.party_tasks = {}
-        self.tasks = {}
-        for k, v in kwargs.items():
-            if hasattr(self, k):
-                setattr(self, k, v)
-
-    def to_dict(self):
-        d = {}
-        for k, v in self.__dict__.items():
-            if v is None:
-                continue
-            d[k] = v
-        return d
-
-    def __str__(self):
-        return str(self.to_dict())
