+++ python/fate_flow/settings.py
@@ -60,7 +60,7 @@ detect_logger = getLogger("fate_flow_detect")
 access_logger = getLogger("fate_flow_access")
 database_logger = getLogger("fate_flow_database")
 
-PARTY_ID = 10000
+PARTY_ID = get_base_config("party_id", "")
 
 MODEL_STORE_PATH = os.path.join(get_fate_flow_directory(), "model")
-DATA_STORE_PATH = os.path.join(get_fate_flow_directory(), "data")
+LOCAL_DATA_STORE_PATH = os.path.join(get_fate_flow_directory(), "data")
